Simplify `mrope_positions` construction - thanks to @imkero for discovering this shortcut.

Co-authored-by: @imkero