@ywang96 2025-01-23
Did some benchmark and the perf diff is pretty minor, but I prefer this cleaner construction

```
main
Request throughput: 1.73 req/s
Total token rate: 6951.79 tok/s

this branch
Request throughput: 1.73 req/s
Total token rate: 6961.40 tok/s
```